+++ src/Module/System/Update/Migration/V7/Migration700008.php
@@ -25,6 +25,7 @@ declare(strict_types=1);
 namespace Ampache\Module\System\Update\Migration\V7;
 
 use Ampache\Config\AmpConfig;
+use Ampache\Module\System\Dba;
 use Ampache\Module\System\Update\Migration\AbstractMigration;
 
 final class Migration700008 extends AbstractMigration
@@ -37,14 +38,16 @@ final class Migration700008 extends AbstractMigration
         $charset   = (AmpConfig::get('database_charset', 'utf8mb4'));
         $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
 
-        $this->updateDatabase('DROP TABLE IF EXISTS `user_playlist_map`;');
-        $this->updateDatabase(
-            sprintf(
-                'CREATE TABLE IF NOT EXISTS `user_playlist_map` (`playlist_id` int(11) UNSIGNED NOT NULL, `user_id` int(11) UNSIGNED NOT NULL, UNIQUE KEY `playlist_user` (`playlist_id`,`user_id`)) ENGINE=%s DEFAULT CHARSET=%s COLLATE=%s ;',
-                $engine,
-                $charset,
-                $collation
-            )
-        );
+        if (!Dba::read('SELECT SUM(`user_id`) FROM `user_playlist_map`;')) {
+            $this->updateDatabase('DROP TABLE IF EXISTS `user_playlist_map`;');
+            $this->updateDatabase(
+                sprintf(
+                    'CREATE TABLE IF NOT EXISTS `user_playlist_map` (`playlist_id` int(11) UNSIGNED NOT NULL, `user_id` int(11) UNSIGNED NOT NULL, UNIQUE KEY `playlist_user` (`playlist_id`,`user_id`)) ENGINE=%s DEFAULT CHARSET=%s COLLATE=%s ;',
+                    $engine,
+                    $charset,
+                    $collation
+                )
+            );
+        }
     }
 }
